app/requirements.txt
Django==2.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djoser==2.0.5
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
